fix(zod): stop alias collision by renaming helper and pin npm import; add duty stubs
