+++ web/src/app/api/reservations/[id]/route.ts
@@ -2,7 +2,7 @@ export const dynamic = 'force-dynamic'
 export const runtime = 'nodejs'
 
 import { NextResponse } from 'next/server'
-import { z } from '@/lib/zod'
+import { z } from '@/lib/zod-helpers'
 import { readUserFromCookie } from '@/lib/auth'
 import { prisma } from '@/src/lib/prisma'
