+++ web/src/app/api/reservations/route.ts
@@ -3,7 +3,7 @@ export const runtime = 'nodejs'
 
 import { NextResponse } from 'next/server'
 import { prisma } from '@/src/lib/prisma'
-import { z } from '@/lib/zod'
+import { z } from '@/lib/zod-helpers'
 import { readUserFromCookie } from '@/lib/auth'
 import type { Prisma } from '@prisma/client'
