+++ web/src/app/groups/[slug]/duties/generate/page.tsx
@@ -1,15 +1,12 @@
-import Link from 'next/link';
-
-export default function GenerateDutiesPage({ params }: { params: { slug: string } }) {
+export default function Page({ params }: { params: { slug: string } }) {
   return (
-    <main className="p-6 space-y-4">
-      <h1 className="text-xl font-semibold">当番を自動割り当て</h1>
+    <main className="max-w-3xl mx-auto p-6 space-y-4">
+      <h1 className="text-2xl font-bold">当番の自動割当</h1>
       <p className="text-sm text-gray-500">グループ: {params.slug}</p>
-      {/* TODO: フォーム実装 */}
-      <div className="text-sm text-gray-400">ここに当番の自動割当ウィザードを実装します。</div>
-      <Link className="text-blue-600 underline" href={`/groups/${params.slug}/duties`}>
-        戻る
-      </Link>
+      {/* 後でウィザード実装 */}
+      <div className="rounded-lg border p-4 bg-white">
+        ここに「期間」「方式（ラウンドロビン/ランダム/手動）」「対象メンバー/曜日」などを置く予定。
+      </div>
     </main>
   );
 }
