+++ web/src/app/groups/[slug]/duties/new-type/page.tsx
@@ -1,14 +1,16 @@
 import Link from 'next/link';
 
-export default function NewDutyTypePage({ params }: { params: { slug: string } }) {
+export default function Page({ params }: { params: { slug: string } }) {
   return (
-    <main className="p-6 space-y-4">
-      <h1 className="text-xl font-semibold">当番タイプを追加</h1>
+    <main className="max-w-3xl mx-auto p-6 space-y-4">
+      <h1 className="text-2xl font-bold">当番タイプの追加</h1>
       <p className="text-sm text-gray-500">グループ: {params.slug}</p>
-      {/* TODO: フォーム実装 */}
-      <div className="text-sm text-gray-400">ここに当番タイプ作成フォームを実装します。</div>
-      <Link className="text-blue-600 underline" href={`/groups/${params.slug}/duties`}>
-        戻る
+      {/* 後でフォーム実装 */}
+      <div className="rounded-lg border p-4 bg-white">
+        ここに「当番タイプ名」「説明」「色」などのフォームを置く予定。
+      </div>
+      <Link href={`/groups/${params.slug}/duties`} className="text-blue-600 underline">
+        ← 当番一覧へ戻る
       </Link>
     </main>
   );
