+++ web/tsconfig.json
@@ -25,9 +25,6 @@
       ],
       "@/src/*": [
         "src/*"
-      ],
-      "zod": [
-        "src/lib/zod"
       ]
     },
     "plugins": [
